+++ src/lsp/main.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { LanguageClient, ServerOptions, LanguageClientOptions, DocumentSymbol, Position } from 'vscode-languageclient/node';
+import { LanguageClient, ServerOptions, LanguageClientOptions, DocumentSymbol, Position, StaticFeature, ClientCapabilities, ServerCapabilities, DocumentSelector } from 'vscode-languageclient/node';
 import * as util from '../utilities'
 import * as config from '../config';
 import { provideClojureDefinition } from '../providers/definition';
@@ -7,11 +7,13 @@ import { provideCompletionItems } from '../providers/completion';
 import { setStateValue, getStateValue } from '../../out/cljs-lib/cljs-lib';
 import { downloadClojureLsp, getLatestVersion } from './download';
 import { readVersionFile, getClojureLspPath } from './utilities';
+import { TestTreeHandler, TestTreeParams } from './types';
 import * as os from 'os';
 import * as path from 'path';
 import * as state from '../state';
 import { provideHover } from '../providers/hover';
 import { provideSignatureHelp } from '../providers/signature';
+import * as cider from '../nrepl/cider'
 
 const LSP_CLIENT_KEY = 'lspClient';
 const RESOLVE_MACRO_AS_COMMAND = 'resolve-macro-as';
@@ -19,12 +21,32 @@ const SERVER_NOT_RUNNING_OR_INITIALIZED_MESSAGE = 'The clojure-lsp server is not
 const lspStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, -1);
 let serverVersion: string;
 
+// The Node LSP client requires the client code to jump through a few hoops in
+// order to enable an experimental feature. This class exists solely to set
+// enable the `experimental.testTree` feature.
+class TestTreeFeature implements StaticFeature {
+
+    initialize(capabilities: ServerCapabilities, documentSelector: DocumentSelector | undefined): void {
+    }
+
+    fillClientCapabilities(capabilities: ClientCapabilities): void {
+        capabilities.experimental = {testTree: true };
+    }
+
+    dispose(): void {
+    }
+}
+
 function createClient(clojureLspPath: string): LanguageClient {
+
     const serverOptions: ServerOptions = {
         run: { command: clojureLspPath },
         debug: { command: clojureLspPath },
+
+
     };
     const clientOptions: LanguageClientOptions = {
+
         documentSelector: [
             { scheme: 'file', language: 'clojure' },
             { scheme: 'jar', language: 'clojure' },
@@ -70,7 +92,7 @@ function createClient(clojureLspPath: string): LanguageClient {
                 try {
                     hover = await provideHover(document, position);
                 } catch (e) {}
-                
+
                 if (hover) {
                     return null;
                 } else {
@@ -299,9 +321,13 @@ function registerEventHandlers(context: vscode.ExtensionContext, client: Languag
     }));
 }
 
-async function startClient(clojureLspPath: string, context: vscode.ExtensionContext): Promise<void> {
+async function startClient(clojureLspPath: string, context: vscode.ExtensionContext, handler: TestTreeHandler): Promise<void> {
     const client = createClient(clojureLspPath);
     console.log('Starting clojure-lsp at', clojureLspPath);
+
+    const testTree : StaticFeature = new TestTreeFeature();
+    client.registerFeature(testTree);
+
     const onReadyPromise = client.onReady();
     lspStatus.text = '$(sync~spin) Initializing Clojure language features via clojure-lsp';
     lspStatus.show();
@@ -314,6 +340,10 @@ async function startClient(clojureLspPath: string, context: vscode.ExtensionCont
     const serverInfo = await getServerInfo(client);
     serverVersion = serverInfo['server-version'];
     sayClientVersionInfo(serverVersion, serverInfo);
+
+    client.onNotification("clojure/textDocument/testTree", (tree: TestTreeParams) => {
+        handler(tree);
+    });
 }
 
 function sayClientVersionInfo(serverVersion: string, serverInfo: any) {
@@ -343,7 +373,7 @@ function registerDiagnosticsCommands(context: vscode.ExtensionContext): void {
     context.subscriptions.push(vscode.commands.registerCommand('calva.diagnostics.openClojureLspLogFile', openLogFile));
 }
 
-async function activate(context: vscode.ExtensionContext): Promise<void> {
+async function activate(context: vscode.ExtensionContext, handler: TestTreeHandler): Promise<void> {
     registerDiagnosticsCommands(context);
     const extensionPath = context.extensionPath;
     const currentVersion = readVersionFile(extensionPath);
@@ -360,7 +390,7 @@ async function activate(context: vscode.ExtensionContext): Promise<void> {
             lspStatus.hide();
         }
     }
-    await startClient(clojureLspPath, context);
+    await startClient(clojureLspPath, context, handler);
 }
 
 function deactivate(): Promise<void> {